Fixing services and controller to REST

+++ src/main/java/com/example/demo/service/UserServiceImpl.java
@@ -1,13 +1,12 @@
 package com.example.demo.service;
 
 
+import com.example.demo.dto.UserDto;
 import com.example.demo.entity.User;
 import com.example.demo.entity.UserRepository;
-import com.example.demo.dto.UserDto;
 import com.example.demo.exception.Exceptions;
 import com.example.demo.exception.ExistingUserException;
 import com.sun.istack.NotNull;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -15,7 +14,6 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 public class UserServiceImpl implements UserService
@@ -28,10 +26,11 @@ public class UserServiceImpl implements UserService
     }
 
     @Override
-    public void createUser(@NotNull UserDto user) throws ExistingUserException
+    public UserDto createUser(@NotNull UserDto user) throws ExistingUserException
     {
         checkIfUserExists(user.getName());
-        userRepository.save(new User(user.getName(), user.getLastName()));
+        User domainUser = new User(user.getName(), user.getLastName());
+        return new UserDto(userRepository.save(domainUser));
     }
 
     private void checkIfUserExists(String name) throws ExistingUserException
@@ -53,18 +52,28 @@ public class UserServiceImpl implements UserService
     @Override
     public void deleteUser(long id)
     {
-        userRepository.deleteById(id);
+        Optional<User> existingUser = userRepository.findById(id);
+
+        if (existingUser.isPresent())
+        {
+            userRepository.deleteById(id);
+        } else
+        {
+            throw new UserNotFoundException(id);
+        }
     }
 
     @Override
     public UserDto getUserById(Long id)
     {
         Optional<User> existingUser = userRepository.findById(id);
-        return existingUser.map(UserDto::new).orElse(null);
+        return existingUser.map(UserDto::new).orElseThrow(() -> new UserNotFoundException(id));
     }
 
     @Override
-    public Page<User> findPaginated(int pageNo, int pageSize, String sortField, String sortDirection) {
+    public Page<User> findPaginated(int pageNo, String sortField, String sortDirection)
+    {
+        final int pageSize = 10;
         Sort sort = sortDirection.equalsIgnoreCase(Sort.Direction.ASC.name()) ? Sort.by(sortField).ascending() :
                 Sort.by(sortField).descending();
